Phase 1: split ann.rs into 5 modules (mod/build/search/select_neighbors/codec)
Surgical refactor with zero behavior change. ann.rs (698 lines) → ann/
directory with each file ≤ 300 lines:

  ann/mod.rs              183 — public API (HnswIndex), Node, Candidate,
                                cosine_dist, tests
  ann/build.rs            222 — build/build_from_*, sample_level, LcgRng
  ann/search.rs           199 — search, attach_vectors, greedy_search,
                                layer_search, ByDistAsc/Desc heap orderings
  ann/select_neighbors.rs 163 — Algorithm 3 (simple, current call site)
                                + Algorithm 4 skeleton (allow(dead_code),
                                wired in Phase 2) + 3 unit tests
  ann/codec.rs            112 — to_bytes, from_bytes, ByteCursor

select_neighbors_simple replaces the inline truncate_and_keep_top calls
in build (build/insert) — same semantics (sort by distance ascending,
truncate to m) plus an explicit id-based tiebreaker for determinism.

97/97 tests green (94 prior + 3 new in select_neighbors::tests).
Clippy clean. fmt clean.

.gitignore: data/measure/corpus_*.nest excluded (per-phase regeneration
artifacts; re-derivable from the LFS-tracked baseline by measure_presets.py).
